general error handling, removed extra logs
